fix(#19): use Decimal for precision

+++ api/services/programa_svc.py
@@ -186,8 +186,8 @@ def get_burnup_custo_programas():
                 continue
             periodo = (row['tempo__ano'], row['tempo__mes'])
             programas = custos_por_periodo.setdefault(periodo, {})
-            info = programas.setdefault(codigo, {'nome': row[campo_nome], 'custo': 0.0})
-            info['custo'] += float(row['custo_periodo'] or 0)
+            info = programas.setdefault(codigo, {'nome': row[campo_nome], 'custo': Decimal(0.0)})
+            info['custo'] += Decimal(row['custo_periodo'] or 0)
 
     acumular(horas_qs, 'programa__codigo_programa', 'programa__nome_programa')
     acumular(compras_qs, 'projeto__programa__codigo_programa', 'projeto__programa__nome_programa')
@@ -197,7 +197,7 @@ def get_burnup_custo_programas():
     for ano, mes in sorted(custos_por_periodo.keys()):
         grupo = {'date_str': f'{mes:02d}/{ano}', 'values': []}
         for codigo, info in custos_por_periodo[(ano, mes)].items():
-            custo_acumulado[codigo] = custo_acumulado.get(codigo, 0.0) + info['custo']
+            custo_acumulado[codigo] = Decimal(custo_acumulado.get(codigo, 0.0)) + info['custo']
             grupo['values'].append({
                 'codigo_programa': codigo,
                 'nome_programa': info['nome'],
